+++ src/ts/interfaces.ts
@@ -79,3 +79,36 @@ export type IDialogueListener<T extends IDialogueNodeType> = (
   oldNodeName: string | null,
   selectedOptionIndex: number
 ) => void;
+
+/**
+ * A simple interface to pass custom HTML renderers to the dialigue logic.
+ */
+export interface IDialogueRenderer {
+  /**
+   * Indicate that the current conversation has ended.
+   * @returns
+   */
+  renderConversationTerminated: () => void;
+
+  /**
+   * Set the text in the question node.
+   * @param {string} questionText - The question text to use.
+   * @param {string|null} npcName - The NPC name to use; if no name was defined this is `null`.
+   */
+  setQuestionText: (questionText: string, npcName: string | undefined) => void;
+
+  /**
+   * Clear the options node. Just for upper level use here.
+   */
+  clearAllOptions: () => void;
+
+  /**
+   * Add a new option node with the given answer text and option index. Use
+   * the option index to send the answer.
+   *
+   * @param {number} answerText - The answer text option.
+   * @param {number} optionIndex - The answer option index in the dialogue structure.
+   * @param {function} onOptionSelected - A callback you must trigger when the option is clicked.
+   */
+  addAnswerOption: (answerText: string, optionIndex: number, onOptionSelected: () => void) => void;
+}